[ios] Treat ECIES error with errSecParam as KeyPermanentlyInvalidated
fix #3763
