+++ app-ios/tutanota/Sources/Keychain/KeychainManager.swift
@@ -1,5 +1,8 @@
 import Foundation
 import LocalAuthentication
+#if !targetEnvironment(simulator)
+import CryptoTokenKit
+#endif
 
 private let TAG = "de.tutao.tutanota.notificationkey."
 private let KEY_PERMANENTLY_INVALIDATED_ERROR_DOMAIN = "de.tutao.tutanota.KeyPermanentlyInvalidatedError"
@@ -9,7 +12,7 @@ class KeyPermanentlyInvalidatedError : TutanotaError {
   init(underlyingError: Error) {
     super.init(message: underlyingError.localizedDescription, underlyingError: underlyingError)
   }
-  
+
   override var name: String {
     get {
       return KEY_PERMANENTLY_INVALIDATED_ERROR_DOMAIN
@@ -21,7 +24,7 @@ class CredentialAuthenticationError : TutanotaError {
   init(underlyingError: Error) {
     super.init(message: underlyingError.localizedDescription, underlyingError: underlyingError)
   }
-  
+
   override var name: String {
     get {
       return CREDENTIAL_AUTHENTICATION_ERROR_DOMAIN
@@ -34,20 +37,20 @@ class KeychainManager : NSObject {
   private static let SYSTEM_PASSWORD_DATA_KEY_ALIAS = "SystemPasswordDataKey"
   private static let BIOMETRICS_DATA_KEY_ALIAS = "BiometricsDataKey"
   private static let DATA_ALGORITHM = SecKeyAlgorithm.eciesEncryptionCofactorVariableIVX963SHA256AESGCM
-  
+
   private let keyGenerator: KeyGenerator
-  
+
   init(keyGenerator: KeyGenerator) {
     self.keyGenerator = keyGenerator
   }
-  
+
   func storeKey(_ key: Data, withId keyId: String) throws {
     let keyTag = self.keyTagFromKeyId(keyId: keyId)
-    
+
     let existingKey = try? self.getKey(keyId: keyId)
-    
+
     let status: OSStatus
-    
+
     if let key = existingKey {
       let updateQuery: [String: Any] = [
         kSecClass as String: kSecClassKey,
@@ -71,7 +74,7 @@ class KeychainManager : NSObject {
       throw TUTErrorFactory.createError("Could not store the key, status: \(status)")
     }
   }
-  
+
   func getKey(keyId: String) throws -> Data? {
     let keyTag = self.keyTagFromKeyId(keyId: keyId)
     let getQuery: [String : Any] = [
@@ -89,7 +92,7 @@ class KeychainManager : NSObject {
       return nil
     }
   }
-  
+
   func removePushIdentifierKeys() throws {
     // TODO: Don't delete all of teh keyz when fingerprints
     let deleteQuery: [String: Any] = [
@@ -100,7 +103,7 @@ class KeychainManager : NSObject {
       throw TUTErrorFactory.createError("Could not delete the keys, status: \(status)")
     }
   }
-  
+
   private func deleteKey(tag: String) throws {
     let deleteQuery: [String : Any] = [
       kSecClass as String: kSecClassKey,
@@ -111,16 +114,16 @@ class KeychainManager : NSObject {
       throw TUTErrorFactory.createError("Failed to delete key: \(status)")
     }
   }
-  
+
   func encryptData(encryptionMode: CredentialEncryptionMode, data: Data) throws -> Data {
     let privateKey = try self.getDataKey(mode: encryptionMode)
     guard let publicKey = SecKeyCopyPublicKey(privateKey) else {
       throw TUTErrorFactory.createError("Cannot get public key from private key for mode \(encryptionMode)")
     }
-    
+
     var error: Unmanaged<CFError>?
     let encryptedData = SecKeyCreateEncryptedData(publicKey, Self.DATA_ALGORITHM, data as CFData, &error) as Data?
-    
+
     guard let encryptedData = encryptedData else {
       switch try self.handleKeychainError(error!, mode: encryptionMode) {
       case .unrecoverable(let error):
@@ -132,13 +135,13 @@ class KeychainManager : NSObject {
     }
     return encryptedData
   }
-  
+
   func decryptData(encryptionMode: CredentialEncryptionMode, encryptedData: Data) throws -> Data {
     let key = try self.getDataKey(mode: encryptionMode)
-    
+
     var error: Unmanaged<CFError>?
     let decryptedData = SecKeyCreateDecryptedData(key, Self.DATA_ALGORITHM, encryptedData as CFData, &error) as Data?
-    
+
     guard let decryptedData = decryptedData else {
       switch try self.handleKeychainError(error!, mode: encryptionMode) {
       case .unrecoverable(let error):
@@ -150,12 +153,12 @@ class KeychainManager : NSObject {
     }
     return decryptedData
   }
-  
+
   private func getDataKey(mode: CredentialEncryptionMode) throws -> SecKey {
     let tag = self.keyAlias(for: mode)
     return try self.fetchDataKey(tag: tag) ?? self.generateDataKey(tag: tag, mode: mode)
   }
-  
+
   private func keyAlias(for encryptionMode: CredentialEncryptionMode) -> String {
     switch encryptionMode {
     case .deviceLock:
@@ -166,7 +169,7 @@ class KeychainManager : NSObject {
       return Self.BIOMETRICS_DATA_KEY_ALIAS
     }
   }
-  
+
   private func fetchDataKey(tag: String) throws -> SecKey? {
     let getQuery: [String : Any] = [
       kSecClass as String: kSecClassKey,
@@ -174,7 +177,7 @@ class KeychainManager : NSObject {
       kSecReturnRef as String: true,
       kSecUseOperationPrompt as String: translate("TutaoUnlockCredentialsAction", default: "Unlock credentials")
     ]
-    
+
     var item: CFTypeRef?
     let status = SecItemCopyMatching(getQuery as CFDictionary, &item)
     switch status {
@@ -186,10 +189,10 @@ class KeychainManager : NSObject {
       throw TUTErrorFactory.createError("Failed to get key \(tag). Status: \(status)")
     }
   }
-  
+
   private func accessControl(for encryptionMethod: CredentialEncryptionMode) -> SecAccessControl {
     let flags: SecAccessControlCreateFlags
-    
+
     switch encryptionMethod {
     case .deviceLock:
         flags = .privateKeyUsage
@@ -198,7 +201,7 @@ class KeychainManager : NSObject {
     case .biometrics:
         flags = [.privateKeyUsage, .biometryCurrentSet]
     }
-    
+
     var error: Unmanaged<CFError>?
     let accessControl = SecAccessControlCreateWithFlags(
       kCFAllocatorDefault,
@@ -206,7 +209,7 @@ class KeychainManager : NSObject {
       flags,
       &error
     )
-    
+
     if let accessControl = accessControl {
       return accessControl
     } else {
@@ -214,20 +217,19 @@ class KeychainManager : NSObject {
       fatalError(error.debugDescription)
     }
   }
-  
+
   private func generateDataKey(tag: String, mode: CredentialEncryptionMode) throws -> SecKey {
     let access = self.accessControl(for: mode)
     return try self.keyGenerator.generateKey(tag: tag, accessControl: access)
   }
-  
+
   private func keyTagFromKeyId(keyId: String) -> Data {
     let keyTag = TAG + keyId
     return keyTag.data(using: .utf8)!
   }
-  
+
   private func handleKeychainError(_ error: Unmanaged<CFError>, mode: CredentialEncryptionMode) throws -> HandledKeychainError {
-    let parsedError = self.keyGenerator.parseKeychainError(error)
-    let message = "Keychain operation failed \(mode)"
+    let parsedError = self.parseKeychainError(error)
     switch parsedError {
     case .authFailure(let error):
       return .unrecoverable(error: CredentialAuthenticationError(underlyingError: error))
@@ -247,10 +249,11 @@ class KeychainManager : NSObject {
       try self.deleteKey(tag: tag)
       return .unrecoverable(error: KeyPermanentlyInvalidatedError(underlyingError: error))
     case .unknown(let error):
+      let message = "Keychain operation failed with mode \(mode)"
       return .unrecoverable(error: TUTErrorFactory.wrapNativeError(withDomain: TUT_ERROR_DOMAIN, message: message, error: error))
     }
   }
-  
+
   private func showPasswordPrompt(reason: String, _ completion: @escaping (Bool?, Error?) -> Void) {
     DispatchQueue.main.async {
       LAContext()
@@ -260,9 +263,44 @@ class KeychainManager : NSObject {
         )
     }
   }
+
+  private func parseKeychainError(_ error: Unmanaged<CFError>) -> KeychainError {
+    let e = error.takeRetainedValue() as Error as NSError
+
+    #if !targetEnvironment(simulator)
+    if #available(iOS 13, *) {
+      if e.domain == TKError.errorDomain && e.code == TKError.Code.corruptedData.rawValue {
+        return .keyPermanentlyInvalidated(error: e)
+      }
+    }
+    #endif
+
+    if e.domain == LAError.errorDomain && e.code == LAError.Code.biometryLockout.rawValue {
+      return .lockout(error: e)
+    }
+
+    if e.domain == LAError.errorDomain {
+      return .authFailure(error: e)
+    }
+
+    // This sometimes happens to some users for unknown reasons
+    if e.domain == NSOSStatusErrorDomain && e.code == Security.errSecParam {
+      return .keyPermanentlyInvalidated(error: e)
+    }
+
+    return .unknown(error: e)
+  }
 }
 
 fileprivate enum HandledKeychainError {
   case recoverable(error: Error)
   case unrecoverable(error: Error)
+}
+
+private enum KeychainError {
+  case authFailure(error: NSError)
+  /// Too many attempts to use biometrics, user must authenticate with password before trying again
+  case lockout(error: NSError)
+  case keyPermanentlyInvalidated(error: NSError)
+  case unknown(error: NSError)
 }
